splitted get_date_stamp
