+++ include/tanja84dk/tools.h
@@ -1,6 +1,5 @@
 #ifndef TANJA84DK_DNSGEN_TOOLS_H
 #define TANJA84DK_DNSGEN_TOOLS_H
-#include <tanja84dk/classes.h>
 
 #include <string>
 #include <vector>
@@ -10,16 +9,7 @@ bool file_exist(const std::string& filename) noexcept;
 
 void save_file(const std::string& filename, const std::vector<std::string>& vector_data) noexcept;
 
-int get_date_stamp() noexcept {
-    time_t timestamp;
-    char output[10];
-    struct tm* datetime;
-    time(&timestamp);
-    datetime = localtime(&timestamp);
-    strftime(output, 10, "%Y%m%d", datetime);
-    std::string data = output;
-    return std::stoi(data);
-}
+int get_date_stamp() noexcept;
 
 }  // namespace Tanja84dk
